Refactor RawRecordsProducer for Row-Wise Processing in raw-records.cpp
Because it produces multiple raw records.
